CORE GAMEPLAY CHANGES: (Mainly for playtest) *add continuous matches and contextual combat* - Replace the round system with a single continuous match flow. - Add Waiting, Countdown, Playing, and Results match phases. - Track replicated chaser time and determine winners or draws. - Restart matches without disconnecting connected players. - Reset player roles, status, timers, and spawn positions between matches. - Start matches only when the required number of players is connected. - Restrict movement, jumping, throwing, and attacks outside active gameplay. - Update the HUD for the new match phases and results. - Improve ragdoll knockback and recover characters at safe ground positions. - Reattach equipped tools correctly on remote clients. - Remove manual tool dropping from the game. - Add separate HandMelee, ToolMelee, and ThrownTool hit types. - Allow the chaser to attack without a tool using shorter hand range. - Prevent runners without tools from performing melee attacks. - Give tool melee and thrown tools distinct stun and knockback values.

+++ Source/Jelly/Private/JellyPlayerState.cpp
@@ -9,7 +9,7 @@ bool AJellyPlayerState::IsChaser() const
 	return bIsChaser;
 }
 
-int32 AJellyPlayerState::GetChaserTime() const
+float AJellyPlayerState::GetChaserTime() const
 {
 	return ChaserTimeSeconds;
 }
@@ -61,6 +61,7 @@ void AJellyPlayerState::GetLifetimeReplicatedProps(TArray<class FLifetimePropert
 	DOREPLIFETIME(AJellyPlayerState, ChaserTimeSeconds);
 	DOREPLIFETIME(AJellyPlayerState, PlayerColorIndex);
 	
+	
 }
 
 void AJellyPlayerState::OnRep_PlayerColorIndex()
@@ -71,5 +72,4 @@ void AJellyPlayerState::OnRep_PlayerColorIndex()
 		JellyCharacter->ApplyPlayerColor();
 	}
 }
-
